+++ test/Application.cs
@@ -1,10 +1,6 @@
 ﻿namespace test
 {
     using Serilog;
-    using System;
-    using System.Collections.Generic;
-    using System.Reflection;
-    using Utilities;
     using Interfaces;
     using Utilities.Interfaces;
     using Examples.Threads;
@@ -29,28 +25,6 @@
             //_initlInitialiseLookups.GeneratePostDeploymentScripts();
             _threading.ThreadLockExample();
             _utility.WaitUserInput();
-        }
-
-        private static void CreateObjectsFromTables(List<Tuple<string, List<Field>>> tables, List<object> tableObjects)
-        {
-            foreach (var table in tables)
-            {
-                var myObject = MyTypeBuilder.CreateNewObject(table.Item1, table.Item2);
-                tableObjects.Add(myObject);
-            }
-        }
-
-        private static void ListObjectPropertiesDetails(object myObject)
-        {
-            IList<PropertyInfo> props = new List<PropertyInfo>(myObject.GetType().GetProperties());
-
-            Console.WriteLine($"Type: {myObject.GetType()}");
-
-            foreach (PropertyInfo prop in props)
-            {
-                object propValue = prop.GetValue(myObject, null);
-                Console.WriteLine($"  Property Name: {prop.Name}\t Property Type: {prop.PropertyType}\n  Property Value: {propValue}");
-            }
         }
     }
 }
